Fixed bug - running ony one analyser
The bug was caused because of misunderstanding of scopes.
When creating new runnable we have no guarantee that
when we run it, all actions performed in it are calling
the same objects that would be called if we would run it
instantly after creating it.

The problem was that I thought that the lambda references
objects themselves and not class variables.